using spf13/cobra for CLI parsing; replacing godeps with dep
