--- Godeps/_workspace/.gitignore
@@ -1,2 +1,0 @@
-/pkg
-/bin
