+++ cmd/merge.go
@@ -1,0 +1,140 @@
+// Copyright © 2019 NAME HERE <EMAIL ADDRESS>
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"log"
+	"os"
+	"path"
+
+	"github.com/cloudfoundry-incubator/candiedyaml"
+	"github.com/mandelsoft/spiff/debug"
+	"github.com/mandelsoft/spiff/dynaml"
+	"github.com/mandelsoft/spiff/flow"
+	"github.com/mandelsoft/spiff/yaml"
+	"github.com/spf13/cobra"
+)
+
+var partial bool
+
+// mergeCmd represents the merge command
+var mergeCmd = &cobra.Command{
+	Use:     "merge",
+	Aliases: []string{"m"},
+	Short:   "Merge stub files into a manifest template",
+	Long:    `Merge a bunch of template files into one manifest, printing it out.`,
+	Args: func(cmd *cobra.Command, args []string) error {
+		if len(args) < 1 {
+			return errors.New("requires at least one arg")
+		}
+		return nil
+	},
+	Run: func(cmd *cobra.Command, args []string) {
+		merge(args[0], partial, args[1:])
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(mergeCmd)
+
+	mergeCmd.Flags().BoolVarP(&debug.DebugFlag, "debug", "d", false, "Print state info")
+
+	mergeCmd.Flags().BoolVarP(&partial, "partial", "p", false, "Allow partial evaluation only")
+}
+
+func merge(templateFilePath string, partial bool, stubFilePaths []string) {
+	var templateFile []byte
+	var err error
+	var stdin = false
+
+	if templateFilePath == "-" {
+		templateFile, err = ioutil.ReadAll(os.Stdin)
+		stdin = true
+	} else {
+		templateFile, err = ReadFile(templateFilePath)
+	}
+
+	if err != nil {
+		log.Fatalln(fmt.Sprintf("error reading template [%s]:", path.Clean(templateFilePath)), err)
+	}
+
+	templateYAMLs, err := yaml.ParseMulti(templateFilePath, templateFile)
+	if err != nil {
+		log.Fatalln(fmt.Sprintf("error parsing template [%s]:", path.Clean(templateFilePath)), err)
+	}
+
+	stubs := []yaml.Node{}
+
+	for _, stubFilePath := range stubFilePaths {
+		var stubFile []byte
+		var err error
+		if stubFilePath == "-" {
+			if stdin {
+				log.Fatalln(fmt.Sprintf("stdin cannot be used twice"))
+			}
+			stubFile, err = ioutil.ReadAll(os.Stdin)
+			stdin = true
+		} else {
+			stubFile, err = ReadFile(stubFilePath)
+		}
+		if err != nil {
+			log.Fatalln(fmt.Sprintf("error reading stub [%s]:", path.Clean(stubFilePath)), err)
+		}
+
+		stubYAML, err := yaml.Parse(stubFilePath, stubFile)
+		if err != nil {
+			log.Fatalln(fmt.Sprintf("error parsing stub [%s]:", path.Clean(stubFilePath)), err)
+		}
+
+		stubs = append(stubs, stubYAML)
+	}
+
+	legend := "\nerror classification:\n" +
+		" *: error in local dynaml expression\n" +
+		" @: dependent of or involved in a cycle\n" +
+		" -: depending on a node with an error"
+
+	prepared, err := flow.PrepareStubs(partial, stubs...)
+	if !partial && err != nil {
+		log.Fatalln("error generating manifest:", err, legend)
+	}
+
+	for no, templateYAML := range templateYAMLs {
+		doc := ""
+		if len(templateYAMLs) > 1 {
+			doc = fmt.Sprintf(" (document %d)", no+1)
+		}
+		if templateYAML.Value() != nil {
+			flowed, err := flow.Apply(templateYAML, prepared)
+			if !partial && err != nil {
+				log.Fatalln(fmt.Sprintf("error generating manifest%s:", doc), err, legend)
+			}
+			if err != nil {
+				flowed = dynaml.ResetUnresolvedNodes(flowed)
+			}
+			yaml, err := candiedyaml.Marshal(flowed)
+			if err != nil {
+				log.Fatalln(fmt.Sprintf("error marshalling manifest%s:", doc), err)
+			}
+			fmt.Println("---")
+			fmt.Println(string(yaml))
+		} else {
+			fmt.Println("---")
+		}
+	}
+}
