+++ cmd/root.go
@@ -1,0 +1,74 @@
+// Copyright © 2019 NAME HERE <EMAIL ADDRESS>
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"os"
+	"strings"
+
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+)
+
+var cfgFile string
+
+// rootCmd represents the base command when called without any subcommands
+var rootCmd = &cobra.Command{
+	Use:     "spiff++",
+	Short:   "deployment manifest toolkit",
+	Version: "1.2.1",
+	Long: `spiff++ is a fork of spiff that provides a compatible extension to spiff based
+on the latest version offering a rich set of new features not yet available in spiff.
+
+spiff is a command line tool and declarative in-domain hybrid YAML templating system.
+While regular templating systems process a template file by substituting the template 
+expressions by values taken from external data sources, in-domain means that the templating 
+engine knows about the syntax and structure of the processed template. It therefore can take
+the values for the template expressions directly from the document processed, including those
+ parts denoted by the template expressions itself.`,
+}
+
+// Execute adds all child commands to the root command and sets flags appropriately.
+// This is called by main.main(). It only needs to happen once to the rootCmd.
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+}
+
+func init() {
+	cobra.OnInitialize(initConfig)
+}
+
+// initConfig reads in config file and ENV variables if set.
+func initConfig() {
+	viper.AutomaticEnv() // read in environment variables that match
+}
+
+func ReadFile(file string) ([]byte, error) {
+	if strings.HasPrefix(file, "http:") || strings.HasPrefix(file, "https:") {
+		response, err := http.Get(file)
+		if err != nil {
+			return nil, fmt.Errorf("error getting [%s]: %s", file, err)
+		}
+		defer response.Body.Close()
+		return ioutil.ReadAll(response.Body)
+	}
+	return ioutil.ReadFile(file)
+}
